Moving existing classes and tests into packages.
