--- src/main/java/org/agileandbeyond/dandd/tddexercise/AbilityType.java
@@ -1,12 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public enum AbilityType {
-
-	STRENGTH,
-	DEXTERITY,
-	CONSTITUTION,
-	WISDOM,
-	INTELLIGENCE,
-	CHARISMA;
-	
-}
