--- src/main/java/org/agileandbeyond/dandd/tddexercise/Alignment.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public enum Alignment {
-	
-	GOOD,
-	EVIL,
-	NEUTRAL
-
-}
