--- src/main/java/org/agileandbeyond/dandd/tddexercise/Armor.java
@@ -1,16 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public abstract class Armor extends Wearable{
-
-	private int armorClassModifier;
-
-	public int getArmorClassModifier() {
-		return armorClassModifier;
-	}
-
-	public void setArmorClassModifier(int modifier) {
-		this.armorClassModifier = modifier;
-	}
-	
-	
-}
