--- src/main/java/org/agileandbeyond/dandd/tddexercise/Boots.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Boots extends Armor {
-
-	public Boots() {
-		this.setPosition(WearablePosition.FEET);
-		this.setArmorClassModifier(1);
-	}
-}
