--- src/main/java/org/agileandbeyond/dandd/tddexercise/BroadSword.java
@@ -1,10 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class BroadSword extends Sword {
-
-	public BroadSword() {
-		this.setAttackModifier(5);
-		this.setDamageModifier(7);
-		this.setTwoHanded(true);
-	}
-}
