--- src/main/java/org/agileandbeyond/dandd/tddexercise/CannotUseShieldException.java
@@ -1,7 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class CannotUseShieldException extends Exception{
-
-	private static final long serialVersionUID = 1L;
-
-}
