--- src/main/java/org/agileandbeyond/dandd/tddexercise/CannotUseWeaponException.java
@@ -1,7 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class CannotUseWeaponException extends Exception {
-
-	private static final long serialVersionUID = 3260671429588045243L;
-
-}
