--- src/main/java/org/agileandbeyond/dandd/tddexercise/Dagger.java
@@ -1,8 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Dagger extends Weapon {
-
-	public Dagger() {
-		this.setTwoHanded(false);
-	}
-}
