--- src/main/java/org/agileandbeyond/dandd/tddexercise/Gauntlets.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Gauntlets extends Armor {
-
-	public Gauntlets() {
-		this.setPosition(WearablePosition.WRISTS);
-		this.setArmorClassModifier(1);
-	}
-}
