--- src/main/java/org/agileandbeyond/dandd/tddexercise/Helmet.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Helmet extends Armor {
-
-	public Helmet() {
-		this.setPosition(WearablePosition.HEAD);
-		this.setArmorClassModifier(1);
-	}
-}
