--- src/main/java/org/agileandbeyond/dandd/tddexercise/MailArmor.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class MailArmor extends Armor {
-
-	public MailArmor() {
-		this.setPosition(WearablePosition.TORSO);
-		this.setArmorClassModifier(4);
-	}
-}
