--- src/main/java/org/agileandbeyond/dandd/tddexercise/RaceType.java
@@ -1,6 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public enum RaceType {
-	HUMAN,
-	ORC
-}
