--- src/main/java/org/agileandbeyond/dandd/tddexercise/Shield.java
@@ -1,9 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Shield extends Armor {
-
-	public Shield(){
-		this.setArmorClassModifier(4);
-		this.setPosition(WearablePosition.SHIELD);
-	}
-}
