--- src/main/java/org/agileandbeyond/dandd/tddexercise/Sword.java
@@ -1,11 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Sword extends Weapon {
-
-	public Sword() {
-		this.setAttackModifier(4);
-		this.setDamageModifier(5);
-		this.setTwoHanded(false);
-	}
-}
-
