--- src/main/java/org/agileandbeyond/dandd/tddexercise/Wearable.java
@@ -1,16 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public abstract class Wearable extends Equipment {
-
-	private WearablePosition position;
-
-	public WearablePosition getPosition() {
-		return position;
-	}
-
-	public void setPosition(WearablePosition position) {
-		this.position = position;
-	}
-	
-	
-}
