--- src/main/java/org/agileandbeyond/dandd/tddexercise/WearablePosition.java
@@ -1,14 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public enum WearablePosition {
-
-	HEAD,
-	NECK,
-	TORSO,
-	WAIST,
-	WRISTS,
-	HANDS,
-	FINGERS,
-	FEET,
-	SHIELD
-}
