+++ src/main/java/org/agileandbeyond/dandd/tddexercise/character/AbilityType.java
@@ -1,0 +1,12 @@
+package org.agileandbeyond.dandd.tddexercise.character;
+
+public enum AbilityType {
+
+	STRENGTH,
+	DEXTERITY,
+	CONSTITUTION,
+	WISDOM,
+	INTELLIGENCE,
+	CHARISMA;
+	
+}
