+++ src/main/java/org/agileandbeyond/dandd/tddexercise/character/Alignment.java
@@ -1,0 +1,9 @@
+package org.agileandbeyond.dandd.tddexercise.character;
+
+public enum Alignment {
+	
+	GOOD,
+	EVIL,
+	NEUTRAL
+
+}
