+++ src/main/java/org/agileandbeyond/dandd/tddexercise/character/Human.java
@@ -1,0 +1,8 @@
+package org.agileandbeyond.dandd.tddexercise.character;
+
+public class Human extends Race {
+
+	public Human() {
+		super();
+	}
+}
