+++ src/main/java/org/agileandbeyond/dandd/tddexercise/character/RaceType.java
@@ -1,0 +1,6 @@
+package org.agileandbeyond.dandd.tddexercise.character;
+
+public enum RaceType {
+	HUMAN,
+	ORC
+}
