+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/Equipment.java
@@ -1,0 +1,33 @@
+package org.agileandbeyond.dandd.tddexercise.equipment;
+
+public abstract class Equipment {
+
+	protected String name;
+	protected int cost;
+	protected int weight;
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public int getCost() {
+		return cost;
+	}
+
+	public void setCost(int cost) {
+		this.cost = cost;
+	}
+
+	public int getWeight() {
+		return weight;
+	}
+
+	public void setWeight(int weight) {
+		this.weight = weight;
+	}
+
+}
