+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/WearablePosition.java
@@ -1,0 +1,14 @@
+package org.agileandbeyond.dandd.tddexercise.equipment;
+
+public enum WearablePosition {
+
+	HEAD,
+	NECK,
+	TORSO,
+	WAIST,
+	WRISTS,
+	HANDS,
+	FINGERS,
+	FEET,
+	SHIELD
+}
