+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/armor/Armor.java
@@ -1,0 +1,18 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.armor;
+
+import org.agileandbeyond.dandd.tddexercise.equipment.Wearable;
+
+public abstract class Armor extends Wearable{
+
+	private int armorClassModifier;
+
+	public int getArmorClassModifier() {
+		return armorClassModifier;
+	}
+
+	public void setArmorClassModifier(int modifier) {
+		this.armorClassModifier = modifier;
+	}
+	
+	
+}
