+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/armor/Gauntlets.java
@@ -1,0 +1,11 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.armor;
+
+import org.agileandbeyond.dandd.tddexercise.equipment.WearablePosition;
+
+public class Gauntlets extends Armor {
+
+	public Gauntlets() {
+		this.setPosition(WearablePosition.WRISTS);
+		this.setArmorClassModifier(1);
+	}
+}
