+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/armor/Helmet.java
@@ -1,0 +1,11 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.armor;
+
+import org.agileandbeyond.dandd.tddexercise.equipment.WearablePosition;
+
+public class Helmet extends Armor {
+
+	public Helmet() {
+		this.setPosition(WearablePosition.HEAD);
+		this.setArmorClassModifier(1);
+	}
+}
