+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/armor/MailArmor.java
@@ -1,0 +1,11 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.armor;
+
+import org.agileandbeyond.dandd.tddexercise.equipment.WearablePosition;
+
+public class MailArmor extends Armor {
+
+	public MailArmor() {
+		this.setPosition(WearablePosition.TORSO);
+		this.setArmorClassModifier(4);
+	}
+}
