+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/armor/Shield.java
@@ -1,0 +1,11 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.armor;
+
+import org.agileandbeyond.dandd.tddexercise.equipment.WearablePosition;
+
+public class Shield extends Armor {
+
+	public Shield(){
+		this.setArmorClassModifier(4);
+		this.setPosition(WearablePosition.SHIELD);
+	}
+}
