+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/weaponry/BroadSword.java
@@ -1,0 +1,10 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.weaponry;
+
+public class BroadSword extends Sword {
+
+	public BroadSword() {
+		this.setAttackModifier(5);
+		this.setDamageModifier(7);
+		this.setTwoHanded(true);
+	}
+}
