+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/weaponry/Dagger.java
@@ -1,0 +1,8 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.weaponry;
+
+public class Dagger extends Weapon {
+
+	public Dagger() {
+		this.setTwoHanded(false);
+	}
+}
