+++ src/main/java/org/agileandbeyond/dandd/tddexercise/equipment/weaponry/Sword.java
@@ -1,0 +1,11 @@
+package org.agileandbeyond.dandd.tddexercise.equipment.weaponry;
+
+public class Sword extends Weapon {
+
+	public Sword() {
+		this.setAttackModifier(4);
+		this.setDamageModifier(5);
+		this.setTwoHanded(false);
+	}
+}
+
