+++ src/main/java/org/agileandbeyond/dandd/tddexercise/exception/CannotUseShieldException.java
@@ -1,0 +1,7 @@
+package org.agileandbeyond.dandd.tddexercise.exception;
+
+public class CannotUseShieldException extends Exception{
+
+	private static final long serialVersionUID = 1L;
+
+}
