+++ src/main/java/org/agileandbeyond/dandd/tddexercise/exception/CannotUseWeaponException.java
@@ -1,0 +1,7 @@
+package org.agileandbeyond.dandd.tddexercise.exception;
+
+public class CannotUseWeaponException extends Exception {
+
+	private static final long serialVersionUID = 3260671429588045243L;
+
+}
