--- src/test/java/org/agileandbeyond/dandd/tddexercise/Human.java
@@ -1,8 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Human extends Race {
-
-	public Human() {
-		super();
-	}
-}
