--- src/test/java/org/agileandbeyond/dandd/tddexercise/Orc.java
@@ -1,13 +1,0 @@
-package org.agileandbeyond.dandd.tddexercise;
-
-public class Orc extends Race {
-	public Orc() {
-		super();
-		this.setRaceType(RaceType.ORC);
-		this.setArmorClassModifier(2);
-		this.setStrengthModifier(2);
-		this.setWisdomModifier(-1);
-		this.setIntelligenceModifier(-1);
-		this.setCharismaModifier(-1);
-	}
-}
